% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estR.R
\name{estR}
\alias{estR}
\title{Estimate latent correlation matrix}
\usage{
estR(
  X1,
  type1,
  X2 = NULL,
  type2 = NULL,
  method = "approx",
  use.nearPD = TRUE,
  nu = 0.01,
  tol = 1e-08,
  verbose = FALSE,
  ratio = 0.9
)
}
\arguments{
\item{X1}{A numeric data matrix (n by p1).}

\item{type1}{A type of variables in \code{X1}, must be one of "continuous", "binary" or "trunc".}

\item{X2}{A numeric data matrix (n by p2).}

\item{type2}{A type of variables in \code{X2}, must be one of "continuous", "binary" or "trunc".}

\item{method}{The calculation method of latent correlation. Either "original" method or "approx". If \code{method = "approx"}, multilinear approximation method is used, which is much faster than the original method. If \code{method = "original"}, optimization of the bridge inverse function is used. The default is "approx".}

\item{use.nearPD}{A logical value indicating whether to use \link[Matrix]{nearPD} or not when the resulting correlation estimator is not positive definite (have at least one negative eigenvalue).}

\item{nu}{Shrinkage parameter for correlation matrix, must be between 0 and 1, the default value is 0.01.}

\item{tol}{Desired accuracy when calculating the solution of bridge function.}

\item{verbose}{If \code{verbose = FALSE}, printing information whether nearPD is used or not is disabled. The default value is FALSE.}

\item{ratio}{The maximum ratio of Kendall's tau and boundary to implement multilinear interpolation.}
}
\value{
\code{estR} returns
\itemize{
      \item{type1: }{Type of the data matrix \code{X1}}
      \item{type2: }{Type of the data matrix \code{X2}}
      \item{R: }{Estimated latent correlation matrix of whole \code{X} = (\code{X1}, \code{X2}) (p1+p2 by p1+p2)}
      \item{R1: }{Estimated latent correlation matrix of \code{X1} (p1 by p1)}
      \item{R2: }{Estimated latent correlation matrix of \code{X2} (p2 by p2)}
      \item{R12: }{Estimated latent correlation matrix between \code{X1} and \code{X2} (p1 by p2)}
}
}
\description{
Estimation of latent correlation matrix from observed data of (possibly) mixed types (continuous/biary/truncated continuous) based on the latent Gaussian copula model.
}
\examples{
### Data setting
n = 1000; p1 = 8; p2 = 5 # sample size and dimensions for two datasets.
rho = .9
# Data generation
simdata = GenData(n=n, type1 = "ternary", type2 = "continuous", p1 = p1, p2 = p2, rho = rho,
                  copula1 = "cube", copula2 = "cube", c1 = c(0, 1), c2 =  NULL)
Sigma = simdata$Sigma; X1 = simdata$X1; X2 = simdata$X2
# Estimate latent correlation matrix with original method
R_nc_org = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "original")$R
# Estimate latent correlation matrix with aprroximation method
R_nc_approx = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "approx")$R

# Heatmap for latent correlation matrix.
Heatmap_R_nc_approx = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "approx")$plotR
}
\references{
Fan J., Liu H., Ning Y. and Zou H. (2017) "High dimensional semiparametric latent graphicalmodel for mixed data" <doi:10.1111/rssb.12168>.

Yoon G., Carroll R.J. and Gaynanova I. (2020) "Sparse semiparametric canonical correlation analysis for data of mixed types" <doi:10.1093/biomet/asaa007>.

Yoon G., Müller C.L., Gaynanova I. (2020) "Fast computation of latent correlations" <arXiv:2006.13875>.
}
