% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Mixed type simulation data generator}
\usage{
GenData(
  n = 100,
  rhos = 0.5,
  copulas = "no",
  types = c("ter", "con"),
  XP = NULL,
  showplot = FALSE
)
}
\arguments{
\item{n}{a positive interger to specify sample size. The default value is 100.}

\item{rhos}{A length of \code{sum(lower.tri(diag(p)))} vector as true latent correlations for between variables.
The default value is 0.5 which means correlations between any two variables are 0.5.}

\item{copulas}{A length of p vector as copula types for each variable, e.g. U = f(Z). Which could be "no", "expo" or "cube".
The default value is "no" which means no copula transformation for any variables.}

\item{types}{A length of p vector as types for each variable. Could be "con" (continuous), "bin" (binary), "tru" (truncated) or "ter" (ternary).
The default value is c("ter", "con") which specifies a pair of variables, the first one is ternary, the second one is continuous.}

\item{XP}{A length of p list of proportions of zeros (and ones for ternary) corresponding to each variable.
 For continuous variable, the corresponding proportion should be NA;
 for binary or truncated variable, the corresponding proportion should be a number between 0 and 1 represents the proportion of zeros;
 for ternary variable, the corresponding proportion should be a pair of numbers between 0 and 1,
 the first number indicates the the proportion of zeros, the second number indicates the proportion of ones.
 The sum of a pair of numbers should between 0 and 1.
The default value is list(c(0.3, 0.5), NA) means the first variable (ternary by default) has 0.3 proportion of zeros and 0.5 proportion of ones,
the second variable (continuous by default) do not need proportion so it's NA.}

\item{showplot}{TRUE or FALSE if you want to plot data if numbers of variables (features) no more than 3. The default value is FALSE, not to generate plot.}
}
\value{
\code{GenData} returns a list containing
\itemize{
      \item{X: }{observed numeric data matrix (n by p).}
      \item{plotX: }{Visualize data matrix X to check if it's consistent with types and latent correlations settings.
                    Histogram of data X if it only has one variable. 2D Scatter plot if it has 2 variables. 3D scatter plot if it has 3 variables.}
}
}
\description{
\code{GenData} is used to generate data of mixed types for the Gaussian copula model. It can generate single, pair or multiple variables.
}
\examples{
# Generate single continuous variable with exponential transformation (always greater than 0)
# and show histogram.
simdata = GenData(n = 100, copulas = "expo", types = "con", showplot = FALSE)
X = simdata$X; plotX = simdata$plotX
# Generate a pair of variables (ternary and continuous) with default proportions
# and without copula transformation.
simdata = GenData()
X = simdata$X
# Generate 3 variables (binary, ternary and truncated)
# corresponding copulas for each variables are "no" (no transformation),
# "cube" (cube transformation) and "cube" (cube transformation).
# binary variable has 30\% of zeros, ternary variable has 20\% of zeros
# and 40\% of ones, truncated variable has 50\% of zeros.
# Then show the 3D scatter plot (data points project on either 0 or 1 on Axis X1;
# on 0, 1 or 2 on Axas X2; on positive domain on Axis X3)
simdata = GenData(n = 100, rhos = c(.3, .4, .5), copulas = c("no", "cube", "cube"),
          types = c("bin", "ter", "tru"), XP = list(.3, c(.2, .4), .5), showplot = TRUE)
X = simdata$X; plotX = simdata$plotX
# Check the proportion of zeros for the binary variable.
sum(simdata$X[ , 1] == 0)
# Check the proportion of zeros and ones for the ternary variable.
sum(simdata$X[ , 2] == 0); sum(simdata$X[ , 2] == 1)
# Check the proportion of zeros for the truncated variable.
sum(simdata$X[ , 3] == 0)
}
