% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallTau.R
\name{Kendall_matrix}
\alias{Kendall_matrix}
\title{Kendall's tau correlation}
\usage{
Kendall_matrix(X, Y = NULL)
}
\arguments{
\item{X}{A numeric matrix (n by p1).}

\item{Y}{A numeric matrix (n by p2).}
}
\value{
\code{Kendall_matrix(X)} returns a p1 by p1 matrix of Kendall's tau correlation coefficients. \code{Kendall_matrix(X, Y)} returns a p1 by p2 matrix of Kendall's tau correlation coefficients.
}
\description{
Calculate Kendall's tau correlation.
\deqn{ \hat{\tau}_{jk} = \frac{2}{n(n-1)}\sum_{1\le i<i'\le n} sign(X_{ji}-X_{ji'}) sign(X_{ki}-X_{ki'}) }
}
