% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{LatentPlot}
\alias{LatentPlot}
\title{Plot latent correlation
\code{LatentPlot} is to plot heatmap for latent correlation matrix.}
\usage{
LatentPlot(
  latentcor,
  xlab = "Variable 1",
  ylab = "Variable 2",
  main = "Latent Correlation for Variable 1 vs. Variable 2"
)
}
\arguments{
\item{latentcor}{latent correlation matrix.}

\item{xlab}{label for horizontal axis.}

\item{ylab}{label for vertical axis.}

\item{main}{main title for graphs.}
}
\value{
\code{LatentPlot} returns a heatmap plot for latent correlation matrix.
}
\description{
Plot latent correlation
\code{LatentPlot} is to plot heatmap for latent correlation matrix.
}
\examples{
### Data setting
n = 1000; p1 = 8; p2 = 5 # sample size and dimensions for two datasets.
rho = .9
# Data generation
simdata = GenData(n=n, type1 = "ternary", type2 = "continuous", p1 = p1, p2 = p2, rho = rho,
                  copula1 = "cube", copula2 = "cube", c1 = c(0, 1), c2 =  NULL)
Sigma = simdata$Sigma; X1 = simdata$X1; X2 = simdata$X2
# Estimate latent correlation matrix with original method
R_nc_org = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "original")$R
# Estimate latent correlation matrix with aprroximation method
R_nc_approx = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "approx")$R


LatentPlot(R_nc_approx)
}
