% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{r_ml_wrapper}
\alias{r_ml_wrapper}
\title{Port function to call multilinear interpolants for continuous developers.}
\usage{
r_ml_wrapper(K, zratio1, zratio2, comb)
}
\arguments{
\item{K}{Kendall's tau, can be provided as a number or a vector for interpolation in batch.}

\item{zratio1}{zratio for first variable. NA for continuous variables, zeros proportions for binary, truncated variables.
For ternary variables, a vector of proportion of zeros as first element, proportions of zeros and ones as second element.
It can be provided as a vector (matrix) for interpolation in batch. See vignettes for detail.}

\item{zratio2}{zratio for second variable. NA for continuous variables, zeros proportions for binary, truncated variables.
For ternary variables, a vector of proportion of zeros as first element, proportions of zeros and ones as second element.
It can be provided as a vector (matrix) for interpolation in batch. See vignettes for detail.}

\item{comb}{Numeric code for types: "10" for binary/continuous; "11" for binary/binary; "20" for truncated/continuous;
"21" for truncated/binary; "22" for truncated/truncated; "30" for ternary/continuous; "31" for ternary/binary;
"32" for ternary/truncated; "33" for ternary/ternary.}
}
\value{
A number or vector of interpolation results.
}
\description{
Port function to call multilinear interpolants for continuous developers.
}
