% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estR.R
\name{estR}
\alias{estR}
\title{Estimate latent correlation matrix}
\usage{
estR(
  X,
  types = c("ter", "con"),
  method = "approx",
  nu = 0.01,
  tol = 1e-08,
  ratio = 0.9,
  corplot = FALSE
)
}
\arguments{
\item{X}{A numeric data matrix (n by p).}

\item{types}{A type of variables in \code{X}, must be one of "continuous", "binary", "trunc" or "ternary".}

\item{method}{The calculation method of latent correlation. Either "original" method or "approx". If \code{method = "approx"}, multilinear approximation method is used, which is much faster than the original method. If \code{method = "original"}, optimization of the bridge inverse function is used. The default is "approx".}

\item{nu}{Shrinkage parameter for correlation matrix, must be between 0 and 1, the default value is 0.01.}

\item{tol}{Desired accuracy when calculating the solution of bridge function.}

\item{ratio}{The maximum ratio of Kendall's tau and boundary to implement multilinear interpolation.}

\item{corplot}{Plot latent correlation matrix \code{R} as a heatmap.}
}
\value{
\code{estR} returns
\itemize{
      \item{K: }{Kendall Tau Matrix of \code{X} (p x p)}
      \item{R: }{Estimated latent correlation matrix of whole \code{X} (p x p)}
}
}
\description{
Estimation of latent correlation matrix from observed data of (possibly) mixed types (continuous/biary/truncated continuous) based on the latent Gaussian copula model.
}
\examples{
# Data generation
X = GenData()
# Estimate latent correlation matrix with original method
R_nc_org = estR(X = X, types = c("tru", "ter"), method = "original")$R
# Estimate latent correlation matrix with aprroximation method
R_nc_approx = estR(X = X, types = c("tru", "ter"), method = "approx")$R

# Heatmap for latent correlation matrix.
Heatmap_R_nc_approx = estR(X = X, types = c("tru", "ter"), method = "approx", corplot = TRUE)$plotR
}
\references{
Fan J., Liu H., Ning Y. and Zou H. (2017) "High dimensional semiparametric latent graphicalmodel for mixed data" <doi:10.1111/rssb.12168>.
Yoon G., Carroll R.J. and Gaynanova I. (2020) "Sparse semiparametric canonical correlation analysis for data of mixed types" <doi:10.1093/biomet/asaa007>.
Yoon G., Müller C.L., Gaynanova I. (2020) "Fast computation of latent correlations" <arXiv:2006.13875>.
}
