% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estR.R
\name{estR}
\alias{estR}
\title{Estimate latent correlation matrix}
\usage{
estR(
  X1,
  type1,
  X2 = NULL,
  type2 = NULL,
  method = "approx",
  use.nearPD = TRUE,
  nu = 0.01,
  tol = 1e-06,
  verbose = FALSE,
  ratio = 0.9
)
}
\arguments{
\item{X1}{A numeric data matrix (n by p1).}

\item{type1}{A type of variables in \code{X1}, must be one of "continuous", "binary" or "trunc".}

\item{X2}{A numeric data matrix (n by p2).}

\item{type2}{A type of variables in \code{X2}, must be one of "continuous", "binary" or "trunc".}

\item{method}{The calculation method of latent correlation. Either "original" method or "approx". If \code{method = "approx"}, multilinear approximation method is used, which is much faster than the original method. If \code{method = "original"}, optimization of the bridge inverse function is used. The default is "approx".}

\item{use.nearPD}{A logical value indicating whether to use \link[Matrix]{nearPD} or not when the resulting correlation estimator is not positive definite (have at least one negative eigenvalue).}

\item{nu}{Shrinkage parameter for correlation matrix, must be between 0 and 1, the default value is 0.01.}

\item{tol}{Desired accuracy when calculating the solution of bridge function.}

\item{verbose}{If \code{verbose = FALSE}, printing information whether nearPD is used or not is disabled. The default value is FALSE.}

\item{ratio}{The maximum ratio of Kendall's tau and boundary to implement multilinear interpolation.}
}
\value{
\code{estR} returns
\itemize{
      \item{type1: }{Type of the data matrix \code{X1}}
      \item{type2: }{Type of the data matrix \code{X2}}
      \item{R: }{Estimated latent correlation matrix of whole \code{X} = (\code{X1}, \code{X2}) (p1+p2 by p1+p2)}
      \item{R1: }{Estimated latent correlation matrix of \code{X1} (p1 by p1)}
      \item{R2: }{Estimated latent correlation matrix of \code{X2} (p2 by p2)}
      \item{R12: }{Estimated latent correlation matrix between \code{X1} and \code{X2} (p1 by p2)}
}
}
\description{
Estimation of latent correlation matrix from observed data of (possibly) mixed types (continuous/biary/truncated continuous) based on the latent Gaussian copula model.
}
\examples{
### Data setting
n = 1000; p1 = 15; p2 = 10 # sample size and dimensions for two datasets.
rho = .9
# Data generation
simdata = GenData(n=n, type1 = "ternary", type2 = "continuous", p1 = p1, p2 = p2, rho = rho,
                  copula1 = "cube", copula2 = "cube", c1 = c(0, 1), c2 =  NULL)
Sigma = simdata$Sigma; X1 = simdata$X1; X2 = simdata$X2
# Estimate latent correlation matrix with original method
R_nc_org = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "original")$R
# Estimate latent correlation matrix with aprroximation method
R_nc_approx = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "approx")$R
# Plot ternary/continuous case estimation via original method.
PlotPair(datapair = cbind(c(Sigma), c(R_nc_org)), namepair = c("Sigma", "R_nc_org"),
         title = "Latent correlation (True vs. Estimated)")
# Plot ternary/continuous case estimation via approximation method.
PlotPair(datapair = cbind(c(Sigma), c(R_nc_approx)), namepair = c("Sigma", "R_nc_approx"),
         title = "Latent correlation (True vs. Estimated)")

cp1 = "exp"; cp2 = "cube"
for (tp1 in c("continuous", "binary", "ternary", "trunc")) {
  for (tp2 in c("continuous", "binary", "ternary", "trunc")) {
    for (md in c("original", "ml", "approx")) {
      c1 = c2 = NULL
      if (tp1 == "binary" | tp1 == "trunc") {
        c1 = 1
      } else if (tp1 == "ternary") {
        c1 = c(1, 2)
      }
      if (tp2 == "binary" | tp2 == "trunc") {
        c2 = 0
      } else if (tp2 == "ternary") {
        c2 = c(0, 1)
      }
      simdata = GenData(n=n, type1 = tp1, type2 = tp2, p1= p1, p2 = p2, rho = rho,
                        copula1 = cp1, copula2 = cp2, c1 = c1, c2 = c2)
      Sigma = simdata$Sigma; X1 = simdata$X1; X2 = simdata$X2
      assign(paste("R", cp1, cp2, tp1, tp2, md, sep = "_"),
             estR(X1 = X1, type1 = tp1, X2 = X2, type2 = tp2, method = md)$R)
      PlotPair(datapair = cbind(c(Sigma), c(get(paste("R", cp1, cp2, tp1, tp2, md, sep = "_")))),
               namepair = c("Sigma", paste("R", cp1, cp2, tp1, tp2, md, sep = "_")),
               title = "Latent correlation (True vs. Estimated)")
    }
  }
}


}
\references{
Fan J., Liu H., Ning Y. and Zou H. (2017) "High dimensional semiparametric latent graphicalmodel for mixed data" <doi:10.1111/rssb.12168>.

Yoon G., Carroll R.J. and Gaynanova I. (2020) "Sparse semiparametric canonical correlation analysis for data of mixed types" <doi:10.1093/biomet/asaa007>.

Yoon G., Müller C.L., Gaynanova I. (2020) "Fast computation of latent correlations" <arXiv:2006.13875>.
}
