% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Mixed type simulation data generator}
\usage{
GenData(n = 100, types = c("tru", "ter"), rhos = 0.5, copulas = c(NA, NA), ...)
}
\arguments{
\item{n}{Sample size}

\item{types}{Types of the  dataset \code{X}. Could be "con", "tru", "bin" or "ter".}

\item{rhos}{True correlation for auto correlated data.}

\item{copulas}{Copula types for the first dataset. U = f(Z), which could be "NA", "exp" or "cube".}
}
\value{
\code{GenData} returns a list containing
\itemize{
      \item{X: }{observed numeric data matrix (n by p).}
}
}
\description{
\code{GenData} is used to generate two sets of data of mixed types for sparse CCA under the Gaussian copula model.
}
\examples{
# Data setting
n = 100; p1 = 15; p2 = 10 # sample size and dimensions for two datasets.
# Data generation
GenData(n = n, type1 = "ternary", type2 = "trunc", p1 = p1, p2 = p2, rho = .9,
        copula1 = "cube", copula2 = "cube", c1 = c(0, 1), c2 = 0)

}
