% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Mixed type simulation data generator}
\usage{
GenData(
  n = 100,
  rhos = 0.5,
  types = c("ter", "con"),
  copulas = c("no", "no"),
  XP = NULL
)
}
\arguments{
\item{n}{Sample size}

\item{rhos}{True correlation for auto correlated data.}

\item{types}{Types of the  dataset \code{X}. Could be "con" (continuous), "bin" (binary), "tru" (truncated) or "ter" (ternary).}

\item{copulas}{Copula types for the first dataset. U = f(Z), which could be "NA", "exp" or "cube".}

\item{XP}{list of proportions corresponding to each data series.}
}
\value{
\code{GenData} returns a list containing
\itemize{
      \item{X: }{observed numeric data matrix (n by p).}
}
}
\description{
\code{GenData} is used to generate two sets of data of mixed types for sparse CCA under the Gaussian copula model.
}
\examples{
GenData()
}
