% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Mixed type simulation data generator
\code{GenData} is used to generate two sets of data of mixed types for sparse CCA under the Gaussian copula model.}
\usage{
GenData(
  n,
  copula1 = "no",
  copula2 = "no",
  type1 = "continuous",
  type2 = "continuous",
  muZ = NULL,
  Sigma,
  c1 = NULL,
  c2 = NULL
)
}
\arguments{
\item{n}{Sample size}

\item{copula1}{Copula type for the first dataset. U1 = f(Z1), which could be either "exp", "cube".}

\item{copula2}{Copula type for the second dataset. U2 = f(Z2), which could be either "exp", "cube".}

\item{type1}{Type of the first dataset \code{X1}. Could be "continuous", "trunc", "binary", "ternary".}

\item{type2}{Type of the second dataset \code{X2}. Could be "continuous", "trunc", "binary", "ternary".}

\item{muZ}{Mean of latent multivariate normal.}

\item{Sigma}{True correlation matrix of latent variable \code{Z1} (p1 by p1).}

\item{c1}{Constant threshold for \code{X1} needed for "trunc", "binary", "ternary" and "ordinal" data type - the default is NULL.}

\item{c2}{Constant threshold for \code{X2} needed for "trunc", "binary", "ternary" and "ordinal" data type - the default is NULL.}
}
\value{
\code{GenData} returns a list containing
\itemize{
      \item{Z1: }{latent numeric data matrix (n by p1).}
      \item{Z2: }{latent numeric data matrix (n by p2).}
      \item{X1: }{observed numeric data matrix (n by p1).}
      \item{X2: }{observed numeric data matrix (n by p2).}
      \item{type: }{a vector containing types of two datasets.}
      \item{c1: }{a matrix of thresholds for \code{X1} for "trunc", "binary", "ternary" and "ordinal" data type.}
      \item{c2: }{a matrix of thresholds for \code{X2} for "trunc", "binary", "ternary" and "ordinal" data type.}
      \item{Sigma: }{true latent correlation matrix of \code{Z1} and \code{Z2} ((p1+p2) by (p1+p2)).}
}
}
\description{
Mixed type simulation data generator
\code{GenData} is used to generate two sets of data of mixed types for sparse CCA under the Gaussian copula model.
}
\examples{
# Data setting
n <- 100; p1 <- 15; p2 <- 10 # sample size and dimensions for two datasets.
# Correlation structure within each data set
set.seed(0)
perm1 <- sample(1:p1+p2, size = p1+p2);
Sigma <- autocor(p1+p2, 0.7)[perm1, perm1]
# blockind <- sample(1:3, size = p2, replace = TRUE);
# Sigma2 <- blockcor(blockind, 0.7)
mu <- rbinom(p1+p2, 1, 0.5)

# Data generation
simdata <- GenData(n=n, copula1 = "exp", copula2 = "cube", type1 = "ternary", type2 = "trunc", muZ = mu, Sigma = Sigma,
                   c1 = matrix(rep(1:2, p1), nrow = 2, ncol = p1), c2 = rep(0, p2))
X1 <- simdata$X1
X2 <- simdata$X2

# Check the range of truncation levels of variables
range(colMeans(X1 == 0))
range(colMeans(X2 == 0))

}
