% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipol.R
\name{interpolation}
\alias{interpolation}
\title{Methods for creating multivariate interpolations on hypercubes.}
\description{
The primary method is \code{\link{ipol}} which
dispatches to some other method.  All the generated
\link{interpolant}s accept as an argument a matrix of column
vectors. The generated functions also accept an argument
\code{threads=getOption('ipol.threads')} to utilize more than
one CPU if a matrix of column vectors is evaluated.  The option
\code{ipol.threads} is initialized from the environment variable
\code{IPOL_THREADS} upon loading of the package. It defaults to \code{1}.
}
\details{
The interpolants are ordinary R-objects and can be saved with \code{save()} and loaded
later with \code{load()} or serialized/unserialized with other tools, just like any R-object.
However, they contain calls to functions in the package, and while the author will make efforts
to ensure that generated interpolants are compatible with future versions of \pkg{ipol},
I can issue no such absolute guarantee.
}
\seealso{
\link{ipol}, \link{interpolant}
}
