% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{PlotPair}
\alias{PlotPair}
\title{Plot true correlation vs estimated correlation
\code{PlotPair} is to check unbiasness of estimation by plotting true correlation from simulation data vs estimated correlation from simulation data.}
\usage{
PlotPair(datapair, namepair = c("X", "Y"), title = "Plot X vs Y")
}
\arguments{
\item{datapair}{matrix for data pairs.}

\item{namepair}{vector for names of data pairs.}

\item{title}{title for graphs.}
}
\value{
\code{PlotPair} returns a plot for data1 against data2 and 45 degree benchmark line.
}
\description{
Plot true correlation vs estimated correlation
\code{PlotPair} is to check unbiasness of estimation by plotting true correlation from simulation data vs estimated correlation from simulation data.
}
\examples{
### Data setting
n = 1000; p1 = 15; p2 = 10 # sample size and dimensions for two datasets.
rho = .9
# Data generation
simdata = GenData(n=n, type1 = "ternary", type2 = "continuous", p1 = p1, p2 = p2, rho = rho,
                  copula1 = "cube", copula2 = "cube", c1 = c(0, 1), c2 =  NULL)
Sigma = simdata$Sigma; X1 = simdata$X1; X2 = simdata$X2
# Estimate latent correlation matrix with original method
R_nc_org = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "original")$R
# Estimate latent correlation matrix with aprroximation method
R_nc_approx = estR(X1 = X1, type1 = "ternary", X2 = X2, type2 = "continuous", method = "approx")$R
# Plot ternary/continuous case estimation via original method.
PlotPair(datapair = cbind(c(Sigma), c(R_nc_org)), namepair = c("Sigma", "R_nc_org"),
         title = "Latent correlation (True vs. Estimated)")
# Plot ternary/continuous case estimation via approximation method.
PlotPair(datapair = cbind(c(Sigma), c(R_nc_approx)), namepair = c("Sigma", "R_nc_approx"),
         title = "Latent correlation (True vs. Estimated)")
PlotPair(datapair = cbind(c(Sigma), c(cor(cbind(X1, X2)))), namepair = c("Sigma", "R_nc_pearson"),
title = "Latent correlation (True) vs. Pearson correlation (Estimated)")

cp1 = "exp"; cp2 = "cube"
for (tp1 in c("continuous", "binary", "ternary", "trunc")) {
  for (tp2 in c("continuous", "binary", "ternary", "trunc")) {
    for (md in c("original", "ml", "approx")) {
      c1 = c2 = NULL
      if (tp1 == "binary" | tp1 == "trunc") {
        c1 = 1
      } else if (tp1 == "ternary") {
        c1 = c(1, 2)
      }
      if (tp2 == "binary" | tp2 == "trunc") {
        c2 = 0
      } else if (tp2 == "ternary") {
        c2 = c(0, 1)
      }
      simdata = GenData(n=n, type1 = tp1, type2 = tp2, p1= p1, p2 = p2, rho = rho,
                        copula1 = cp1, copula2 = cp2, c1 = c1, c2 = c2)
      Sigma = simdata$Sigma; X1 = simdata$X1; X2 = simdata$X2
      assign(paste("R", cp1, cp2, tp1, tp2, md, sep = "_"),
             estR(X1 = X1, type1 = tp1, X2 = X2, type2 = tp2, method = md)$R)
      PlotPair(datapair = cbind(c(Sigma), c(get(paste("R", cp1, cp2, tp1, tp2, md, sep = "_")))),
               namepair = c("Sigma", paste("R", cp1, cp2, tp1, tp2, md, sep = "_")),
               title = "Latent correlation (True vs. Estimated)")
    }
  }
}


}
