% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{autocor}
\alias{autocor}
\title{Construct a correlation matrix
Functions to create autocorrelation matrix (p by p) with parameter rho and block correlation matrix (p by p) using group index (of length p) and (possibly) different parameter rho for each group.}
\usage{
autocor(p, rho)
}
\arguments{
\item{p}{Specified matrix dimension.}

\item{rho}{Correlation value(s), must be between -0.99 and 0.99. Should be a scalar for \code{autocor}, and either a scalar or a vector of the same length as the maximal \code{blockind} K for \code{blockcor}.}
}
\value{
Correlation matrix \code{Sigma}
}
\description{
Construct a correlation matrix
Functions to create autocorrelation matrix (p by p) with parameter rho and block correlation matrix (p by p) using group index (of length p) and (possibly) different parameter rho for each group.
}
